Second commit
Continued work on the toString-method in the MLP class and added the
ability to write the MLP to a file. Also added cost function
functionality and expected output to the MLP.

TODO: Maybe add a vector/matrix class to make the code a bit more readable instead of using double arrays. Also Start work on implementing the back propagation to the network.